build: Cleanup/fix usage of ALL_PREBUILT
While working with AOSP latest branch, we found that, the
variable has changed and causing build failures.
To get rid of build failures and to be in sync with AOSP Build
System, we are deprecating the use of the variable.

As the NOTICE file is already in the repository, commenting
the ALL_PREBUILT usage.

Change-Id: Id4ccc4093eafc271c27e9dd1a786844517d58300
CRs-Fixed: 314291