Add script to generate the build configuration.
On a standard linux install, the configure script will generate header
files enabling/disabling features based on the system's configuration.

On android, we don't want to use the host configuration and need to wrap
that call with the right build parameters (CFLAGS, LDFLAGS, etc) in
order to generate the right configuration.
Note: this script should be run everytime we update curl.

BUG: 22347561

Change-Id: Icd44dbd35daa811e16f327fc4ec23f0cbe1d8780